backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Steven Hollingsworth <shollingsworth@barracuda.com>"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26.119"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
boto3-stubs = {extras = ["dynamodb", "ssm", "apigatewaymanagementapi"], version = "^1.26.119"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
